Add the Kotlin counterpart to the Java [FizzBuzzSequence] implementation.
It also re-uses the unit-test from the Java implementation but it
does not implement the Java interface [FizzBuzzSequence]. The reason for
that is that Kotlin works best with collection classes directly but
does not really need the vararg version of "translate" because a list
of Int is really quickly written. Also it works rather awkward with
Java streams which would have to be converted into a Kotlin collection
first. On the other hand does the Java implementation not need
a version of "translate" that accept a List<Integer> because the
vararg-version and the IntStream version look much better when called.

The interface [FizzBuzzSequence] is still needed to re-use the Java
test-cases but is adapted by the [FizzBuzzSequenceKotlinTest]
through an object.