Add a Scala implementation of the word-wrap kata plus a Scalatest.
This implementation uses a different approach than WordWrapJava. It
takes chunks of the column-width from the beginning of the text until
the text is completely consumed. It looks for the rightmost space in each
chunk and splits there to avoid splits inside a word. No split line
will ever start or end with a space. This is not feature really
but workaround for the way chunks are processed. See comment in on "wrap".

This implementation is backed by Scalatest. I would liked to have reused
the Spock specification for this implementation too but that turned out
to be harder than expected: the Scala collection classes confuse Groovy.
And even when wrapped with the JavaConversions utilities, simple
equality based matching of list does not work. Plus: the Scala and the
Java implementation behave differently regarding multiple subsequent
spaces inside a line: the Java implementation removes them, while the
Scala implementation does not.