Hamcrest matcher "IsMonthlyRepayment" can be set up to match individual properties of "MonthlyRepayment" instances.
Before, five property values had to be provided. This had the drawback
that it was not clear which value referred to which property,
especially since four of them were of type Money. Now, each property
is named as its property which makes the assertion much more readable.
It is also possible now to match only against a subset of
the five supported properties.